Added configuration file using ConfigObj library in Python.
Added configspec file that contains information regarding the format of the data in config.ini.

On startup, the configspec.ini file is loaded and then the config.ini file is validated against the configspec. If the data fails to validate, then the errors are reported, an exception is thrown and the program will quit. This config variable is then sent to the MainWindow where it can be used to retrieve configuration values.

Currently, the only config values present are the scanner ports for the primary/secondary scanner.