Begin Work On Switching Between Windows
Changed favorite's menu and main menu from QMainWindow to a QWidget. These will be placed within a QStackedWidget on a QMainWindow which will allow for easy traversal between the windows.

Added CentralWidget QMainWindow that contains a stacked widget with all of the menus. The config and barcode scanners are setup here and the appropiate signals are routed whenever the stacked widget changes its current index.

Switched components that had the background color set to the same as the style to a transparent background color. This produces the same results but makes our lives easier if the background color of the program to change.

Added functionality to back and home button in favorite's menu.