Finished QStackedWidget implementation. Can easily switch between widgets now
Corrected SmartShop home logo at the top. When switching between main window and the favorite's window, it would slightly shift from the left to the right because they weren't in the exact same location.

Set current tab index to the favorite's tab in favorite's window.

Delete all tabs besides favorite's in favorite's window. Each tab will be added programmatically by querying the database.

Rename items in favorite's window to more appropiate names since the GUI is primarily finished for the favorite's window.

Previously, I explored the idea of having a global config variable for the config.ini so that any class can access. This was located in the constants module. I decided to remove that and pass it down to each window that needs it.