Create GetCategory and GetCategories function in DatabaseManager for retrieving the category information.
Switched psycopg2 cursor to DictCursor which allows you to reference data from fetch in a dict with column name. For example, executeState['columnName'] is valid instead of having to do the index of the column which is quite arbitrary. Currently, I am only using the dict reference within favorite's menu, but I will switch the other references to dict instead. This is beneficial because the column ID may change over time if a new column is inserted while referencing by a column name will always be the same. In addition, the code looks cleaner when you reference by a column name.

To the windows contained in the stacked widget, I now pass essential objects to the class such as database manager, config, and central window.