Began Implementation of Manual Add Dialog
Updated ManualAddDialog window filenames to adhere to camelCase standard (manualAddDialog)

Edited manualAddDialog GUI. Renamed items in design, changed font sizes, and removed quantity field.

Created TouchButton class which inherits from QPushButton. It overrides the mousePressEvent and mouseReleaseEvent functions and determines if the button is pressed (left-clicked) or long-pressed (right-clicked). The signal clicked is overloaded with two boolean inputs: one for the checked state and the other for whether it was long pressed or not.

Promoted buttons in manualAddDialog to TouchButton class to allow for regular and long pressing implementation.

Created TouchSpinBox button which inherits from QSpinBox. It implements scrolling effects by handling QScrollPrepareEvent and QScrollEvent. It handles left and right-clicking of the spin box for kinetic scrolling.