Added checkbox to ManualAddDialog that will add the item to the favorite's menu.
Created TouchCheckbox class to allow for left and right-clicking of the checkbox.

Added GetFavoritesCount to DatabaseManager class that gets the largest favorite index count. This is used when the "Add Item to Favorites" checkbox is checked in the ManualAddDialog so that it sets the favorite index to the last in the list.