Added SqlTableModel which mimics QSqlTableModel that uses a SQL database to query the rows.
QSqlTableModel requires that their SQL module be used and provides no support for the Python module psycopg2. SqlTableModel is a model that utilizes a psycopg2 connection to supply data to the table. The model currently provides read-only access to the data and no provisions have been made to edit the data in the table.

Replaced simple string list model for each tab with the SqlTableModel class that was made. Each tab now appropriately loads the correct items.