Add/Remove Button on Favorite's Menu is Disabled/Enabled based on Highlighted Items
By default, the add/remove buttons are enabled on the favorite's menu in the constructor. From there, a signal is setup to check whether an item has been selected on any list(the presumption is that only the ACTIVE list can be selected at a time, so all of the lists are hooked up to the same function). The second signal checks for a tab change on the widget and adjusts the buttons being enabled or disabled from there.