Added SqlTileTableModel which shows items in Favorite's Tab of Fav Menu
I created the SqlTileTableModel which does essentially the same thing as SqlTableModel except it displays it as tiles instead of rows. For example, in SqlTileTableModel with columnsPerRow = 5, row 1 and column 1 would be linear index 6 in the SQL records (with an assumed column of 0 meaning the first field retrieved is the name). On the other hand, SqlTableModel would consider row 1 and column 1 to be the second data record in SQL query and the second column.

Altered both SqlTableModel and SqlTileTableModel to remove any uses of QVariant since Python handles variables like variants anyway.

Implemented QTileViewModel for favorite's model.