Only valid indices can be selected or are enabled in SqlTileTableModel
So, in a table you must have the same number of columns and rows for the entire table. If you have columnsPerRow = 5 and 32 elements total, that means the last row will have three unused items that MUST be there. This is because the last row cannot have a column count of 2, it must be the same as everything else (5 in this case).

This was handled by overriding the flags function in SqlTileTableModel that says any index that is not valid (does not have a record) is not selectable and is not enabled.