Add option for custom queries in SqlTableModel since I had to do some subquery for the purchase history selection. (I had to retrieve the name from the inventory table with the given item ids and get the other data in the purchase_history table)
Add column alignment list to SqlTableModel which sets the text alignment of a column. This is so the qty and date purchased columns will be center aligned in the purchase history table to even out the 10px padding (10px padding with left-aligned text means its all to the right).

Add CSS to stylesheet to prevent blue outline on QListView and QTableView when active.