Created query that will select a list of suggested items. The query favors items that match at the beginning of the item name versus in the middle. It also favors items that match with the appropiate casing but it will show items that do not have the right case.
Added clear button to virtual keyboard to allow for quickly clearing the text if you select a suggested item.

Selecting an item from the list puts the name in the text edit.

Add HideItems boolean to SqlTableModel which will have the model show NO items if this is set true. If you dont want the model to show the query results, then this should be set true.

Add limit clause to the SqlTableModel. This MUST go after the ORDER BY clause and so I cannot insert it into the custom query.