Added usage_history table to SQL database which is exactly the same structure as purchase_history table but represents the items per day the user has thrown away.
Added column avg_usage_rate that is a real (A.K.A. float) for the average number of items the user throws out per day.

Create timer that will tick each day at a specified time to check the usage times

Get the new entries in usage_history and calculate new avg_usage_rate's for the corresponding items.