Added updatePurchaseHistory function to DatabaseManager that inserts item into purchase_history table
The function is given an item number, date, and qty that was purchased. From there, the query will search for a row based on item ID and purchase date. If it already exists, the item is incremented by qty, otherwise it is inserted as qty.

Added basic test to make sure the query works as expected