Update mainWindow shopping lists when shown
If the inventory data in the database has changed since the window was left, we will update the models so that the changes can be seen. Also, resize the columns in case the quantity text is larger than the current size. Finally, set the sorting for the lists back to default.