Setup combobox's appropiately in all of the dialog boxes
Added line in constructor of parent of combobox to make the view not track mouse movements. When using touchscreen, this prevents the first time you start scrolling to select an item. Instead, you have to just select it and not move the scrollbar while doing it.

In all of the dialog boxes, the combobox's were missing the ScrollPerPixel to make it not scroll ridiculously fast as well.