Remove slot for check in/out button since it does nothing. The checked state is directly checked when an item is scanned in or out.
Remove hasSelection debug printout since the favorite's window is finished.

Instead of initializing the expiration box once at initialization, I moved it to the DisplayExpirationBox function. It is a tradeoff between memory vs speed but in reality either method doesnt matter. It makes more sense to just construct it when it is needed and then dispose of it.

Fixed some instances of where a string was checked against being != ''. Instead, a cleaner looking way is to say if str because it will check if it is empty OR if it equals None.

Update favorite's window so that it will update the categories tab each time it is shown. New categories can be added via the NewItemDetails box and so it should check whenever it is first shown