Instead of initializing the new item details dialog box once at initialization, I moved it to the AddItemToInventory function. It is a tradeoff between memory vs speed but in reality either method doesnt matter. It makes more sense to just construct it when it is needed and then dispose of it.
Within the category combo box in manualAddDialog and newItemDetails, I switched to using the user variant data to identify categories better.

QComboBox contains a findText function to search for a text within it. Switched to using that instead of a for loop which reinvents the wheel.

For order_index entry in database, it is not necessarily true that the indices are in order. A category could become deleted and thus it is not safe to do that. Instead, I switched the query to get the largest order_index and set the next to one plus that when adding new categories.

Added custom exception hook because when handling signals and events in PyQt5 it does not trigger exceptions. This is why you will frequently just see the program crash and have no idea why. Adding a custom exception hook fixes this as well as I can print the information to a log with logging module in case a console is not available.

Update favorite's window favorite tab whenever it is being shown.

Began work on "scan to cancel" functionality for the expiration box. Some experimentation will need to be done to see how this can practically work on the Pi.