Update TouchLineEdit to allow for auto complete models within it
Modify GetItemFromInventory in DatabaseManager to return None if the item was not found

The item name in NewItemDetails menu is now configured to suggest existing item names. Thus, you can link multiple UPCs to the same item ID in the database.

In NewItemDetails, when the item text changes, it will check if the item name already exists in the database. And if so, it will link the UPC to the existing item. It will also set the category and favorite's checkbox to whatever the existing item has and then disable the controls since they are not configurable. Otherwise, if the item name is new, it will enable these controls since they can be configured.

Since the name has to be unique for each item, this means a check needs to be done whenever a new item is entered. A warning label was added to the manualAddDialog box since this adds a new item to the table. Implemented functionality and prevent user from submitting item when it already exists in the database.

Fixed bug where starting virtual keyboard with text in it would select all of the text. I don't think we want that.

Add check for a valid name in new item dialog box. That way the user does not enter an empty string and cause that to be an item

Fixed bug where category combo box wasnt set properly if the name was initially an item that already exists