Move checkOutBox from being constructed once at the initialization of BarcodeManager to being constructed whenever it is required to pop up. This also eliminates the need for a ResetToDefault function.
Alter DecrementQuantityForItem in DatabaseManager to make the SQL query itself limit the quantity to zero. This was done using an if statement in Python but doing it in the PostgreSQL statement doesnt hurt either.

Fixed bug where RemoveFromInventory did not factor in the quantity per package variable

Added scan to cancel functionality to the check out box. If the user scans an item out whilst the box is open, it will automatically accept whatever changes are present on the box.

Fixed combo box to be touch-friendly on check out box. The scrolling is more smooth and doesn't select an item while scrolling.