Fixed bug where program crashes if you select pending category in new item details box.
In the new item details box, there is a pending category that has an asterisk next to its name. This category is only added to the database if you select the item to have that category. However, the program would crash if you did so.

Also, added icon files. Forgot to commit them.